+++ get_connected_users_5_munin.php
@@ -100,46 +100,44 @@ function get_connections($bd, $sub, $last_connect_minutes)
         //check the existence of configuration.php
         $config_file = '';
         if (is_file($bd.'/'.$dir.$sub.'/app/config/configuration.php')) {
-            // Chamilo 1.10
+            // Chamilo 1.10+
             $config_file = $bd.'/'.$dir.$sub.'/app/config/configuration.php';
         } elseif (is_file($bd.'/'.$dir.$sub.'/main/inc/conf/configuration.php')) {
             // Chamilo 1.9
             $config_file = $bd.'/'.$dir.$sub.'/main/inc/conf/configuration.php';
         }
         if (!empty($config_file)) {
+            $_configuration = [];
             $inc = include_once($config_file);
-            $dbh = mysqli_connect($_configuration['db_host'], $_configuration['db_user'],
-                $_configuration['db_password']);
+            $dsn = 'mysql:dbname='.$_configuration['main_database'].';host='.$_configuration['db_host'];
+            try {
+                $dbh = new PDO($dsn, $_configuration['db_user'], $_configuration['db_password']);
+            } catch (PDOException $e) {
+                die('Failed to connect to database: '.$e->getMessage());
+            }
             if ($inc !== false && $dbh !== false) {
-                $db = $_configuration['main_database'];
                 $sql = "SELECT CONCAT(UTC_DATE(),' ',UTC_TIME())";
-                $res = mysqli_query($dbh, $sql);
-                $row = mysqli_fetch_row($res);
+                $res = $dbh->query($sql);
+                $row = $res->fetch();
                 $current_date = $row[0];
-                $track_online_table = $db.'.track_e_online';
+                $track_online_table = 'track_e_online';
                 $query = "SELECT count(login_user_id) ".
                     " FROM ".$track_online_table.
                     " WHERE DATE_ADD(login_date, ".
                     "INTERVAL $last_connect_minutes MINUTE) >= '".$current_date."'  ";
-                //echo $query."\n";
-                $res = mysqli_query($dbh, $query);
+                $res = $dbh->query($query);
                 if ($res === false) {
                     $num = 0;
-                    //echo "          There was a query error for the following portal\n";
                 } else {
-                    $row = mysqli_fetch_row($res);
+                    $row = $res->fetch();
                     $num = $row[0];
                 }
-                //echo sprintf("[%7d]",$num)." users connected to ".$_configuration['root_web']." last $last_connect_minutes'\n";
                 $cut_point = 7;
                 if (substr($_configuration['root_web'], 0, 5) == 'https') {
                     $cut_point = 8;
                 }
                 $connections[str_replace('.', '_', substr($_configuration['root_web'], $cut_point, -1))] = $num;
                 $match_count += $num;
-                mysqli_close($dbh);
-            } else {
-                //echo "$bd/$dir$sub:could not open configuration.php or database:\n";
             }
         }
     }